introduce a helper class for the simulator setup
