+++ sim/quic-network-simulator-helper.h
@@ -1,0 +1,18 @@
+#ifndef QUIC_NETWORK_SIMULATOR_HELPER_H
+#define QUIC_NETWORK_SIMULATOR_HELPER_H
+
+#include "ns3/node.h"
+
+using namespace ns3;
+
+class QuicNetworkSimulatorHelper {
+public:
+  QuicNetworkSimulatorHelper();
+  Ptr<Node> GetLeftNode() const;
+  Ptr<Node> GetRightNode() const;
+
+private:
+  Ptr<Node> left_node_, right_node_;
+};
+
+#endif /* QUIC_NETWORK_SIMULATOR_HELPER_H */
